rm unused lines
