+++ frontend/__tests__/routes/_oh.test.tsx
@@ -35,12 +35,6 @@ describe("frontend/routes/_oh", () => {
   });
 
   it("should render", async () => {
-    const getConfigSpy = vi.spyOn(OpenHands, "getConfig");
-    getConfigSpy.mockResolvedValue({
-      APP_MODE: "saas",
-      GITHUB_CLIENT_ID: "test-id",
-      POSTHOG_CLIENT_KEY: "test-key",
-    });
     renderWithProviders(<RouteStub />);
     await screen.findByTestId("root-layout");
   });
@@ -53,13 +47,6 @@ describe("frontend/routes/_oh", () => {
   });
 
   it("should not render the AI config modal if the settings are up-to-date", async () => {
-    const getConfigSpy = vi.spyOn(OpenHands, "getConfig");
-    getConfigSpy.mockResolvedValue({
-      APP_MODE: "saas",
-      GITHUB_CLIENT_ID: "test-id",
-      POSTHOG_CLIENT_KEY: "test-key",
-    });
-
     settingsAreUpToDateMock.mockReturnValue(true);
     renderWithProviders(<RouteStub />);
 
@@ -120,13 +107,6 @@ describe("frontend/routes/_oh", () => {
   });
 
   it("should not render the user consent form if the user has already made a decision", async () => {
-    const getConfigSpy = vi.spyOn(OpenHands, "getConfig");
-    getConfigSpy.mockResolvedValue({
-      APP_MODE: "saas",
-      GITHUB_CLIENT_ID: "test-id",
-      POSTHOG_CLIENT_KEY: "test-key",
-    });
-
     localStorage.setItem("analytics-consent", "true");
     renderWithProviders(<RouteStub />);
